Created the classes we'll need
