+++ model/context_site.rb
@@ -1,0 +1,3 @@
+class ContextSite < SnapshotCluster
+
+end
