+++ model/context_site_ad.rb
@@ -1,0 +1,3 @@
+class ContextSiteAd < SnapshotCluster
+
+end
