+++ model/context_site_ad_snapshot.rb
@@ -1,0 +1,3 @@
+class ContextSiteAdSnapshot < Snapshot
+
+end
