+++ model/yt_search.rb
@@ -1,0 +1,3 @@
+class YtSearch < SnapshotCluster
+
+end
