+++ model/yt_side_ad.rb
@@ -1,0 +1,3 @@
+class YtSideAd < SnapshotCluster
+
+end
