+++ model/yt_side_ad_snapshot.rb
@@ -1,0 +1,3 @@
+class YtSideAdSnapshot < Snapshot
+
+end
