+++ model/yt_video_ad.rb
@@ -1,0 +1,3 @@
+class YtVideoAd < SnapshotCluster
+
+end
