+++ model/yt_video_ad_snapshot.rb
@@ -1,0 +1,3 @@
+class YtVideoAdSnapshot < Snapshot
+
+end
